Adding error handling
The clean_and_preprocess_data function is now wrapped in a try...except block to catch any OSError exceptions related to accessing the input or output folders.
Inside the loop that iterates over files, the pd.read_csv operation is wrapped in a try...except block to catch two types of exceptions:
pd.errors.EmptyDataError: This exception is raised when the CSV file is empty. In this case, a message is printed, and the script moves on to the next file.
pd.errors.ParserError: This exception is raised when there is an error parsing the CSV file (e.g., incorrect format, invalid data). In this case, a message is printed with the exception details, and the script moves on to the next file.
The cleaned_data.to_csv operation is also wrapped in a try...except block to catch two types of exceptions:
PermissionError: This exception is raised when the script does not have permission to write the output file. In this case, a message is printed indicating the permission error.
Exception: This catches any other exceptions that might occur while saving the file. A message is printed with the exception details.